+++ src/java/beans/ManageDriversBean.java
@@ -71,22 +71,52 @@ public class ManageDriversBean implements Serializable {
     }
 
     public void deleteSelectedDriver() {
+        boolean flag = driverDao.checkBusesDrivers(selectedDriver.getDriverId());
+
         try {
-            driverDao.deleteDriver(selectedDriver.getDriverId());
-            sessionBean.navigateManageDrivers();
+
+            if (flag) {
+                error_message_header = "Error!";
+                error_message_content = "Please unassign this driver from bus before deleting it.";
+                RequestContext.getCurrentInstance().showMessageInDialog(new FacesMessage(FacesMessage.SEVERITY_INFO, error_message_header, error_message_content));
+            } else {
+                driverDao.deleteDriver(selectedDriver.getDriverId());
+                sessionBean.navigateManageDrivers();
+            }
 
         } catch (Exception ex) {
             error_message_header = "Error!";
             error_message_content = ex.getMessage();
-              if(error_message_content.contains("ORA-02292: integrity constraint (BUSES.BUSES_DRIVERS_FK2) violated - child record found")){
-                error_message_content="This driver is currently assigned to a bus. Please unassign this driver from bus first then try again.";
-            
+            if (error_message_content.contains("ORA-02292: integrity constraint (BUSES.BUSES_DRIVERS_FK2) violated - child record found")) {
+                error_message_content = "This driver is currently assigned to a bus. Please unassign this driver from bus first then try again.";
+
             }
-            
 
             RequestContext.getCurrentInstance().showMessageInDialog(new FacesMessage(FacesMessage.SEVERITY_INFO, error_message_header, error_message_content));
             Logger.getLogger(ManageDriversBean.class.getName()).log(Level.SEVERE, null, ex);
         }
+    }
+
+    public boolean checklBusesDrivers() {
+        boolean flag = driverDao.checkBusesDrivers(selectedDriver.getDriverId());
+
+        try {
+            if (flag) {
+                error_message_header = "Error!";
+                error_message_content = "Please unassign this driver from bus before editing it.";
+                RequestContext.getCurrentInstance().showMessageInDialog(new FacesMessage(FacesMessage.SEVERITY_INFO, error_message_header, error_message_content));
+            } else {
+                sessionBean.navigateAddEditDriver();
+            }
+
+        } catch (Exception ex) {
+            error_message_header = "Error!";
+            error_message_content = ex.getMessage();
+
+            RequestContext.getCurrentInstance().showMessageInDialog(new FacesMessage(FacesMessage.SEVERITY_INFO, error_message_header, error_message_content));
+            Logger.getLogger(ManageRouteStopsBean.class.getName()).log(Level.SEVERE, null, ex);
+        }
+        return flag;
     }
 
 }
